+++ res/readysignal.js
@@ -78,11 +78,24 @@ function detectAndPrepareReadinessComponents() {
             console.log(SIGNAL_MATHJAX_READY_FOR_EXPORT);
         } else {
             // Formulas found. Wait for MathJax to process them.
-            window.MathJax.Hub.Queue(function () {
-                window.MoodleQuizArchiver.readySignals.mathjax = true;
-                console.log(SIGNAL_MATHJAX_READY_FOR_EXPORT);
-            });
-            window.MathJax.Hub.processSectionDelay = 0;
+            if (window.MathJax.version.startsWith('2')) {
+                console.debug("MathJax v2 detected. Waiting for MathJax to process equations ...");
+                window.MathJax.Hub.Queue(function () {
+                    window.MoodleQuizArchiver.readySignals.mathjax = true;
+                    console.log(SIGNAL_MATHJAX_READY_FOR_EXPORT);
+                });
+                window.MathJax.Hub.processSectionDelay = 0;
+            } else if (window.MathJax.version.startsWith('3')) {
+                console.debug("MathJax v3 detected. Waiting for MathJax to process equations ...");
+                window.MathJax.startup.promise.then(() => {
+                    window.MoodleQuizArchiver.readySignals.mathjax = true;
+                    console.log(SIGNAL_MATHJAX_READY_FOR_EXPORT);
+                });
+            } else {
+                console.error("Unknown MathJax version detected: " + window.MathJax.version);
+                console.debug("Just waiting 3 seconds ...")
+                setTimeout(() => {console.log(SIGNAL_MATHJAX_READY_FOR_EXPORT);}, 3000);
+            }
         }
     } else {
         console.log(SIGNAL_MATHJAX_NOT_FOUND);