+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version X.Y.Z (YYYY-MM-DD)
+## Version 3.Y.Z (YYYY-MM-DD)
 
 - Add attempt folder name customization by variable path field in API
 - Pass job UUID to `generate_attempt_report` and `get_attempts_metadata` Moodle webservice functions