+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support for Moodle 5.0 / MathJax 3.x
 - Add attempt folder name customization by variable path field in API
 - Pass job UUID to `generate_attempt_report` and `get_attempts_metadata` Moodle webservice functions
+- Update Python dependencies
 
 
 ## Version 2.3.0 (2025-04-01)